Move components from component directory to their own dirs. Delete useless msw package.
