+++ frontend/smart-kickers-game/src/components/Game/GameResults.css
