+++ frontend/smart-kickers-game/src/components/Game/GameResults.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import { Button } from '../components/Button';
-import { pointsManipulation } from '../apis/goalManipulation.js';
+import { Button } from '../Button/Button';
+import { pointsManipulation } from '../../apis/goalManipulation.js';
 import './GameResults.css';
-import { TeamID, ScoreChange } from '../constants/score.js';
+import { TeamID, ScoreChange } from '../../constants/score.js';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
 function GameResults({ blueScore, whiteScore }) {
