+++ frontend/smart-kickers-game/src/components/Game/GameStatistics.css
